Make tapping on cover art to open playlist configurable
This introduces an option to control whether the current playlist is
displayed when the album cover is tapped on the Now Playing screen. This
option is enabled by default, so the previous behavior is not changed.
Disabling this option makes it easier to access the fading button panel
at the bottom of the cover.