Fix offline scrobbling for servers without lucene support Current implementation of offline scrobbling is to save all required data to submit scrobble in local shared preferences file and when back online, DSub sends them to server.
When in Offline mode, dsub uses pathes to local files as their identifies
as opposite to Online mode where server side identifies are utilized.
In order to send scrobble it is required to specify server side ID
in a query.

Thus, Dsub has to match identifies of what it has scrobbled while offline
(paths to files in file system) to server side identifies. To do so, it relies
on
1. Querying local database, which contains mapping local path to file <-> server side id.
It sounds good but we don't live in perfect world and there are couple flaws in this implementation:
a) Dsub's local database only contains information for songs that have been played in online
mode at least once.
b) It just doesn't work since in scrobble OfflineMusicService.scrobble method id of a song comes with ".complete"
postfix and in database it's saved without this ".complete" postfix so it fails to lookup server side id.

2. Querying the server with lucene syntax search clause. Subsonic api doesn't say anything about lucene
search syntax support and not all servers have support for it (i.e. see this issue for navidrome
https://github.com/navidrome/navidrome/issues/1466). Besides, it isn't precise in case you have multiple
songs with same artist and title (i.e. multiple editions of an album).

This commit and according PR focus on fixing problems with looking up local path to file <-> server side id.
It is achieved with removing ".complete" substring from all identifies which
comes to OfflineMusicSerivce.scrobble. And by adding all cached songs to local db on download stage
so when DSub switches to offline mode it always has server side identifies for all media files it has
locally.