WRITE_EXTERNAL_STORAGE permission logic fix
Multiple permissions cannot be requested at once, they can be given in String[] or via the callback of the previous requested permission, that is the method used here. Also changed the API level of the request to 28 and below to match the AndroidManifest and documentation involving scoped storage and MediaStore.